moved exceptions to mylibrelab package
